Add missing blocks for overloading
This blocks extremly important for the new custom-theme-handle-strategy
of twig.
There should no longer be a child theme (like in Smarty), but all
overrides should take place in one module. In order to overload APEX
assets, the preload area of the CSS/JS/FONTS files must be overloadable
in addition to the actual loading of the CSS/JS/FONTS files.

Signed-off-by: Anton Fedurtsya <anton@fedurtsya.com>